Update README.md
Lakshmana Pandian P's portfolio website is a visually appealing showcase of his skills and accomplishments as a Software Developer and Python Developer. The design exudes a professional yet creative atmosphere, featuring a clean layout with a cohesive color scheme and modern typography.

The header section introduces Lakshmana with a prominent name, professional title, and location. The navigation bar, with its smooth transition effects and hover animations, guides visitors seamlessly through various sections of the portfolio.

Sections such as "Summary," "Objective," and "Education" provide comprehensive insights into Lakshmana's academic and professional journey. Each section is not only well-structured but also visually engaging, with a subtle fade-in animation on scroll.

The project section highlights key works with brief descriptions and links to GitHub repositories. Certifications and experiences are presented with clarity, accompanied by relevant details. The skills section showcases Lakshmana's proficiency in various programming languages and tools.

Awards and honors are displayed, emphasizing Lakshmana's achievements in internships and hackathons. The footer, featuring a clean design, includes a copyright notice and adds a finishing touch to the website.

The website is not only functional but also interactive, with smooth scrolling, active section highlighting, and animated section entrances. Overall, Lakshmana Pandian P's portfolio is a compelling presentation of his skills, experiences, and personality in the field of software development.